🐛 Migration for column `tg_id` type

+++ app/infrastructure/database/models/user.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from app.infrastructure.database import Base
-from sqlalchemy import Boolean, Column, DateTime, Integer, String
+from sqlalchemy import BigInteger, Boolean, Column, DateTime, Integer, String
 from sqlalchemy.orm import relationship
 
 
@@ -9,7 +9,7 @@ class UserModel(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True)
-    tg_id = Column(Integer, unique=True, nullable=False)
+    tg_id = Column(BigInteger, unique=True, nullable=False)
 
     language_code = Column(String, nullable=True)
     show_nsfw = Column(Boolean, default=False, nullable=False)
